fix cgi bug
- In Cgi.cpp, God bless me for not storing the result from replaceAll
  back to 'field'. Took many hours of debugging to catch this.

+++ Response.cpp
@@ -14,7 +14,7 @@ using std::map;
 
 std::string my_itoa(size_t i);
 
-int my_atoi(std::string s);
+ssize_t my_atoi(std::string s);
 
 int my_hex_to_int(std::string s);
